Added ShakeIt browser to Dial button.
Moved all ShakeIt specific JavaScript code, which is used by both Hotkey and Dial,
to a separate file and class.

Issue #48
