+++ StreamDeckSimHub.Plugin/StreamDeckSimHub.Plugin.csproj
@@ -46,6 +46,9 @@
       <None Update="pi\hotkey.html">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       </None>
+      <None Update="pi\js\shakeit.js">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </None>
       <None Update="pi\js\hotkey.js">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       </None>
