+++ StreamDeckSimHub.Plugin/pi/hotkey.html
@@ -76,6 +76,7 @@
 <script src="libs/js/api.js"></script>
 <script src="libs/js/property-inspector.js"></script>
 <script src="libs/js/utils.js"></script>
+<script src="js/shakeit.js"></script>
 <script src="js/hotkey.js"></script>
 
 </body>
